requirements.txt
Flask==3.1.0
google-generativeai==0.8.4
torch==2.6.0
transformers==4.49.0